+++ packages/bruno-app/src/providers/App/ConfirmAppClose/index.js
@@ -1,5 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
+import { findCollectionByUid, flattenItems, isItemARequest } from 'utils/collections';
+import each from 'lodash/each';
+import filter from 'lodash/filter';
+import groupBy from 'lodash/groupBy';
 import SaveRequestsModal from './SaveRequestsModal';
 import { isElectron } from 'utils/common/platform';
 import { completeQuitFlow } from 'providers/ReduxStore/slices/app';
@@ -8,6 +13,8 @@ const ConfirmAppClose = () => {
   const { ipcRenderer } = window;
   const [showConfirmClose, setShowConfirmClose] = useState(false);
   const dispatch = useDispatch();
+  const collections = useSelector((state) => state.collections.collections);
+  const tabs = useSelector((state) => state.tabs.tabs);
 
   useEffect(() => {
     if (!isElectron()) {
@@ -27,7 +34,26 @@ const ConfirmAppClose = () => {
     return null;
   }
 
-  return <SaveRequestsModal onConfirm={completeQuitFlow} onClose={() => setShowConfirmClose(false)} />;
+  const getAllDraftRequests = () => {
+    const draftRequests = [];
+    const tabsByCollection = groupBy(tabs, (t) => t.collectionUid);
+    Object.keys(tabsByCollection).forEach((collectionUid) => {
+      const collection = findCollectionByUid(collections, collectionUid);
+      if (collection) {
+        const items = flattenItems(collection.items);
+        const drafts = filter(items, (item) => isItemARequest(item) && item.draft);
+        each(drafts, (draft) => {
+          draftRequests.push({
+            ...draft,
+            collectionUid: collectionUid
+          });
+        });
+      }
+    });
+    return draftRequests
+  }
+
+  return <SaveRequestsModal draftRequests={getAllDraftRequests()} onConfirm={completeQuitFlow} onClose={() => setShowConfirmClose(false)} />;
 };
 
 export default ConfirmAppClose;
